requirements.txt
gunicorn
deepface
Flask
numpy
opencv_python
Pillow
requests
tensorflow
gevent